+++ models/users.py
@@ -1,0 +1,14 @@
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+
+class Users(db.Model):
+    id = db.Column(db.Integer,primary_key = True)
+    name = db.Column(db.String)
+    email = db.Column(db.String,unique = True)
+    password = db.Column(db.String)
+
+    def __init__(self,name,email,password) -> None:
+        self.name = name
+        self.email = email
+        self.password = password
